DB에서 Farmer값 안 넘어오는 issue 해결
