+++ lib/firebase/database_provider.dart
@@ -18,17 +18,17 @@ class DatabaseProvider with Transfomer {
 //        .transform(transSensorData);
 //  }
 
-//  // documents 모두 갖고옴
-//  Stream<List<SensorData>> linkSensorUUID() {
-//    return _firestore
-//        .collection(COLLECTION_NAME)
-//        .snapshots()
-//        .transform(transSensorUUID);
-//  }
-
-  Stream<Farmer> linkSensorData(String farmerKey) {
+  // documents 모두 갖고옴
+  Stream<List<SensorData>> linkSensorUUID() {
     return _firestore
         .collection(COLLECTION_NAME)
+        .snapshots()
+        .transform(transSensorUUID);
+  }
+
+  Stream<Farmer> linkFarmerData(String farmerKey) {
+    return _firestore
+        .collection(COLLECTION_FARMER)
         .document(farmerKey)
         .snapshots()
         .transform(transFarmerData);
@@ -40,6 +40,32 @@ class DatabaseProvider with Transfomer {
     final DocumentSnapshot snapshot = await farmerRef.get();
     return _firestore.runTransaction((Transaction tx) async {
       if (!snapshot.exists) await tx.set(farmerRef, Farmer.createMap(email));
+    });
+  }
+
+  Future<Map<String, dynamic>> addSensorUUID(
+      {String farmerKey, String sensorUUID}) async {
+    final DocumentReference farmerRef =
+        _firestore.collection(COLLECTION_FARMER).document(farmerKey);
+    final DocumentSnapshot snapshot = await farmerRef.get();
+
+    return _firestore.runTransaction((Transaction tx) async {
+      if (snapshot.exists) await tx.update(farmerRef, <String, dynamic>{
+        KEY_SENSOR_UUID : FieldValue.arrayUnion([sensorUUID])
+      });
+    });
+  }
+
+  Future<Map<String, dynamic>> delSensorUUID(
+      {String farmerKey, String sensorUUID}) async {
+    final DocumentReference farmerRef =
+    _firestore.collection(COLLECTION_FARMER).document(farmerKey);
+    final DocumentSnapshot snapshot = await farmerRef.get();
+
+    return _firestore.runTransaction((Transaction tx) async {
+      if (snapshot.exists) await tx.update(farmerRef, <String, dynamic>{
+        KEY_SENSOR_UUID : FieldValue.arrayRemove([sensorUUID])
+      });
     });
   }
 }
