+++ lib/main.dart
@@ -1,46 +1,67 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:smartfarm/firebase/db_data/mine_sensor_data.dart';
+import 'package:smartfarm/firebase/database_provider.dart';
+import 'package:smartfarm/firebase/db_data/provider/mine_farmer_data.dart';
 import 'package:smartfarm/screen/info_page.dart';
-import 'package:smartfarm/screen/auth_page.dart';
 import 'package:smartfarm/screen/sensor_list_page.dart';
-
+import 'package:smartfarm/screen/sign_in.dart';
 import 'package:smartfarm/utils/progress_indicator.dart';
 import 'screen/info_page.dart';
 
-void main() => runApp(ChangeNotifierProvider<MineSensorData>(
-    create: (context) => MineSensorData(), child: MyApp()));
+void main() => runApp(ChangeNotifierProvider<MineFarmerData>(
+    create: (context) => MineFarmerData(), child: MyApp()));
 
-bool login_Checker = true;
-
+bool isItFirstData = true;
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      title: 'SmartFarmer',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      //home: InfoPage(),
-
       home: StreamBuilder<FirebaseUser>(
-          stream: FirebaseAuth.instance.onAuthStateChanged,
-          //User가 로그인한 경우 Stream을 통해서 <FirebaseUser>정보를 준다. 로그아웃 된 상태라면 Stream을 통해 NULL값을 준다.
-          builder: (context, snapshot) {
-            if (login_Checker) {
-              login_Checker = false;
-              return Progress_Indicator();
-            } else {
-              if (snapshot.hasData) {
-                //데이터가 있다는 것은 유저가 있다는 것.
-                return InfoPage();
-              }
-              return AuthPage();
+        stream: FirebaseAuth.instance.onAuthStateChanged,
+        builder: (context, snapshot){
+          if(isItFirstData){
+            isItFirstData = false;
+            return Progress_Indicator();
+          } else{
+            if(snapshot.hasData){
+              databaseProvider.createFarmer(farmerKey: snapshot.data.uid, email: snapshot.data.email);
+              var myUserData = Provider.of<MineFarmerData>(context);
+              databaseProvider.linkFarmerData(snapshot.data.uid).listen((event) {
+                myUserData.setFarmerData(event);
+              });
+              return InfoPage();
             }
-          }),
+            return SignIn();
+          }
+        }),
+
+
+
+//      Consumer<MineFarmerData>(
+//        builder: (context, farmerData, child) {
+//          switch (farmerData.status) {
+//            case MineFarmerStatus.progress:
+//              FirebaseAuth.instance.currentUser().then((firebaseFarmer) {
+//                if (firebaseFarmer == null)
+//                  farmerData.setFarmerStatus(MineFarmerStatus.none);
+//                else
+//                  databaseProvider
+//                      .linkFarmerData(firebaseFarmer.uid)
+//                      .listen((farmer) {
+//                    farmerData.setFarmerData(farmer);
+//                  });
+//              });
+//              return Progress_Indicator();
+//            case MineFarmerStatus.exist:
+//              return InfoPage();
+//
+//            default:
+//              return SignIn();
+//          }
+//        },
+//      ),
     );
   }
 }
